requirement.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
python-multipart
streamlit
PyPDF2
python-docx
pandas
pytesseract
pdf2image
